+++ src/index.js
@@ -2,7 +2,12 @@ import React from "react";
 import ReactDOM from "react-dom";
 
 import App from "./components/App";
-
+import { GameProvider } from "./components/GameContext";
 const rootElement = document.getElementById("root");
 
-ReactDOM.render(<App />, rootElement);
+ReactDOM.render(
+  <GameProvider>
+    <App />
+  </GameProvider>,
+  rootElement
+);
